Move source into src folder so 'combPackage' folder is treated as a package
More on this here: https://web.cs.wpi.edu/~cs2102/b12/Assignments/making-packages.html

but the short story is that each Java class ought to be in a folder
matching its package name (so `package foo.bar` corresponds to folders
`./foo/bar`) -- which we already were doing in the project. The second
thing is that all the source code belongs in one folder which is not the
package folder. Some Java implementations (e.g. IntelliJ Idea IDE) won't
recognize the top level directory as a package.