Rename package from combPackage to combinatorics
Justification: the "Package" in combPackage is redundant -- we don't
write

    Integer heightInteger = 42;
    String nameString = "egg salad";
    class SetClass { /* ... */ }

for the same reason.

If we drop the "Package", "comb" seems too short. For most code, we read
it more often than we write it so it's more time efficient for the
humans involved if we optimize for reading. We do that by avoiding
abbreviations. The longer a named thing exists, the more important this
is -- so a package name affects the entire project, while an iteration
counter affects only the few lines inside the iteration. (But the longer
your for loop, the more important it is to name the counter variable
something meaningful.)