Introduces packages for functions, operations, and sets
It is conventional to have one file for each class that will be used in
another file. The current pattern of having a single file for each of
functions, operations, and test makes it difficult to test those
packages independently, impossible to use them from another package, and
(possibly most importantly) leads to very long files.

Long files slow down development because:
 - multiple people editing the same long file at the same time are more
   likely to cause [merge conflicts](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/about-merge-conflicts)
   when integrating their code together
 - IDEs have good support for switching back and forth between different
   files while working on multiple (e.g. looking at the base `Operation`
   while writing a new one). The support for jumping up and down within
   one file is not nearly as good, so it leads to the electronic
   equivalent of rifling through a stack of papers trying to find the one
   you were just looking at

Next steps would be to update the README to change "add to the bottom of
the file" to "add a file to the package folder" and maybe extract the
template into another file (?) or consider if the format ought to change
in response to the new package structure.