Got things started using the Element interface idea.
I think I'm on the right track, and I suspect this will drastically simplify the Operation functionality.  I got rid of the the input setters, operators, and type checking in Operation base, and I'm pretty sure I can get rid of the input fields in Operation subclasses too.

Since I changed CombinatorialSet to be an arraylist of Elements, I need to fix a lot of things before it will compile again.  I haven't put in the time yet to get unit testing working, so going to just overhaul everything so that can compile from main.